pyproject.toml
[project]
name = "handson"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.85",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
]
